requirements.txt
jinja2==3.1.4
schemax-openapi==0.0.4
PyYAML==6.0.2
requests==2.32.3
